Moved PrivateRoute to contexts folder
